Move main logic into internal cli
